Refactor: update 메서드 병합
