+++ src/main/java/com/sprint/mission/discodeit/entity/Channel.java
@@ -1,0 +1,96 @@
+package com.sprint.mission.discodeit.entity;
+
+import java.io.Serializable;
+import java.time.Instant;
+import java.util.UUID;
+
+public class Channel implements Serializable {
+
+    private final UUID id;
+    private final long createdAt;
+    private long updatedAt;
+
+    private String name;
+    private String description;
+    private ChannelType type;
+
+    public Channel(String name, String description, ChannelType type) {
+
+        this.id = UUID.randomUUID();
+        this.createdAt = Instant.now().getEpochSecond();
+
+        this.name = name;
+        this.description = description;
+        this.type = type;
+    }
+
+    public static boolean validation(String name, String description){
+        if (name == null || name.isEmpty()) {
+            return false;
+        }
+        if (description == null || description.isEmpty()) {
+            return false;
+        }
+        return true;
+    }
+
+    public UUID getId() {
+        return id;
+    }
+
+    public long getCreatedAt() {
+        return createdAt;
+    }
+
+    public long getUpdatedAt() {
+        return updatedAt;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public ChannelType getType() {
+        return type;
+    }
+
+    public void setUpdatedAt() {
+        this.updatedAt = Instant.now().getEpochSecond();
+    }
+
+    public void setChannel(String name, String description, ChannelType type){
+        if(name !=null && !name.equals(this.name)){
+            this.name = name;
+        }else{
+            throw new IllegalArgumentException("입력한 채널이름: "+name+"이 기존 값과 같습니다.");
+        }
+
+        if(description !=null && !description.equals(this.description)){
+            this.description = description;
+        } else{
+            throw new IllegalArgumentException("입력한 설명: "+description+"이 기존 값과 같습니다.");
+        }
+
+        if(type != null && !type.equals(this.type)){
+            this.type = type;
+        }
+
+        setUpdatedAt();
+    }
+
+    @Override
+    public String toString() {
+        return "Channel{" +
+                "id=" + id +
+                ", createdAt=" + createdAt +
+                ", updatedAt=" + updatedAt +
+                ", name='" + name + '\'' +
+                ", description='" + description + '\'' +
+                ", type=" + type +
+                '}';
+    }
+}
