+++ src/main/java/com/sprint/mission/discodeit/entity/ChannelType.java
@@ -1,0 +1,6 @@
+package com.sprint.mission.discodeit.entity;
+
+public enum ChannelType {
+    PUBLIC,
+    PRIVATE
+}
