+++ src/main/java/com/sprint/mission/discodeit/service/ChannelService.java
@@ -1,0 +1,28 @@
+package com.sprint.mission.discodeit.service;
+
+import com.sprint.mission.discodeit.entity.Channel;
+import com.sprint.mission.discodeit.entity.ChannelType;
+
+
+import java.util.List;
+import java.util.UUID;
+
+public interface ChannelService {
+
+/*
+    //db 로직
+    UUID save(Channel channel);
+    Channel findOne(UUID id);
+    List<Channel> findAll();
+    UUID update(Channel channel);
+    UUID delete (UUID id);
+*/
+
+    //서비스 로직
+    UUID create(String name, String description, ChannelType type);
+    Channel read(UUID id);
+    List<Channel> readAll();
+    Channel updateChannel(UUID id, String name, String description, ChannelType type);
+    UUID deleteChannel(UUID id);
+}
+
