--- src/main/java/com/srint/mission/discodeit/entity/ChannelType.java
@@ -1,6 +1,0 @@
-package com.srint.mission.discodeit.entity;
-
-public enum ChannelType {
-    PUBLIC,
-    PRIVATE
-}
