--- src/main/java/com/srint/mission/discodeit/repository/UserRepository.java
@@ -1,15 +1,0 @@
-package com.srint.mission.discodeit.repository;
-
-import com.srint.mission.discodeit.entity.User;
-
-import java.util.List;
-import java.util.UUID;
-
-public interface UserRepository {
-
-    UUID save(User user);
-    User findOne(UUID id);
-    List<User> findAll();
-    UUID update(User user);
-    UUID delete(UUID id);
-}
